pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["doudouu0504 <selinafs880504@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
psycopg = {extras = ["binary"], version = "^3.2.3"}
django-allauth = "^65.3.0"
requests = "^2.32.3"
pillow = "^11.0.0"
boto3 = "^1.35.81"
django-storages = "^1.14.4"
python-dotenv = "^1.0.1"
django-simple-captcha = "^0.6.0"
pyjwt = "^2.10.1"
cryptography = "^44.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
